feat(painel-obras): detalhes inline + planejamento da semana
Converte o modal "Registro do dia" em uma linha expansível dentro da
própria tabela do Painel de Obras, alinhada ao design do restante do
site e com foco em planejamento semanal.

UX
- Ícone de chevron (ChevronRight → rotaciona 90° ao abrir) como
  primeiro elemento da célula Cliente/Obra; substitui o botão
  "Registro do dia" dos ações.
- Row expandida renderiza DailyLogInline ocupando todas as colunas
  (colSpan = PAINEL_COLUMN_COUNT), com fundo sutil (accent/10) e
  borda superior/inferior para reforçar o contexto.
- Hover da row principal ganha tom mais forte quando expandida.
- Cores migradas para tokens do site (text-[hsl(var(--primary))])
  tanto na célula Cliente/Obra quanto nos botões de ação.

DailyLogInline (novo)
- Substitui DailyLogDialog (removido). Não é mais um Dialog — renderiza
  um <div> para uso dentro de <TableCell colSpan>.
- Nova seção "Planejamento da semana" persistida dentro de `notes`,
  separada por marcador `---PLAN---` (evita nova coluna no BD).
- Data default passa a ser a segunda-feira da semana corrente
  (startOfWeek weekStartsOn: 1) — registro de semana.
- Mantém seções Serviços em execução, Prestadores no local e
  Observações gerais, todas usando os tokens de cor do site.

PainelObras.tsx
- Remove estado/render do modal; adiciona `expandedIds: Set<string>`
  e `toggleExpanded`.
- ObraRow agora recebe `expanded` + `onToggleExpanded` (em vez de
  `onOpenDailyLog`) e retorna `<>...<TableRow expanded>...</>`.
- Adiciona constante `PAINEL_COLUMN_COUNT = 12`.
- Remove import BookOpen e o botão correspondente na coluna de ações.
